botify/requirements.txt

Flask==2.2.5
Flask-RESTful==0.3.9
Flask-And-Redis==1.0.0
python-json-logger==2.0.2
mmh3==3.0.0
gevent
gunicorn
scikit-learn
Werkzeug==2.2.2
pandas
joblib
numpy